added Recipe execution with countdown timer, small bug once the timer has stopped it cant be resumed

+++ src/main/java/org/hua/cookhelp/Recipe.java
@@ -116,7 +116,7 @@ public class Recipe {
                 String timeUnit = matcher.group(2);
             
                 try {
-                    double timeDuration = Double.parseDouble(timeStr);
+                    long timeDuration = Long.parseLong(timeStr);
                     Step step = new Step(s, timeDuration, timeUnit);
                     addStep(step);
                 } catch (NumberFormatException e) {
